DS-1261: Extract method related to environment variable to its own class following SRP
